Adding flask to dockerfile.
