+++ mltrace/server/Dockerfile
@@ -1,0 +1,23 @@
+# pull official base image
+FROM python:3.8.1-slim-buster
+
+# set work directory
+WORKDIR /src
+
+# set environment variables
+ENV PYTHONDONTWRITEBYTECODE 1
+ENV GIT_PYTHON_REFRESH=quiet
+
+# install dependencies
+RUN pip install --upgrade pip
+COPY requirements.txt requirements.txt
+RUN pip install -r requirements.txt
+
+# expose port
+EXPOSE 5000
+
+# copy project
+COPY . .
+
+# run command
+CMD [ "python", "./mltrace/server/manage.py", "run", "--host", "0.0.0.0" ]
