+++ mltrace/server/manage.py
@@ -1,0 +1,10 @@
+from flask.cli import FlaskGroup
+
+from mltrace.server import app
+
+
+cli = FlaskGroup(app)
+
+
+if __name__ == "__main__":
+    cli()
