+++ requirements.txt
@@ -1,0 +1,28 @@
+attrs==20.3.0
+autopep8==1.5.6
+click==7.1.2
+Flask==1.1.2
+gitdb==4.0.7
+GitPython==3.1.14
+greenlet==1.0.0
+iniconfig==1.1.1
+itsdangerous==1.1.0
+Jinja2==2.11.3
+joblib==1.0.1
+MarkupSafe==1.1.1
+packaging==20.9
+pluggy==0.13.1
+psycopg2-binary==2.8.6
+py==1.10.0
+pycodestyle==2.7.0
+pyparsing==2.4.7
+pytest==6.2.2
+python-dateutil==2.8.1
+python-dotenv==0.16.0
+pytz==2021.1
+six==1.15.0
+smmap==4.0.0
+SQLAlchemy==1.4.3
+threadpoolctl==2.1.0
+toml==0.10.2
+Werkzeug==1.0.1
